pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ISHWorkTools"
description = "Make working with ICM 7.10 easier"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["automate", "work", "setup"]
authors = [
  { name = "Johannes Metzner", email = "jm@jometzner.de" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11"
]
dependencies = [
  "gitdb==4.0.10",
  "gitpython==3.1.41",
  "jinja2==3.1.2",
  "markupsafe==2.1.2",
  "pyyaml==6.0.2",
  "setuptools-scm>=8.1.0",
  "setuptools>=75.8.0",
  "six==1.16.0",
  "smmap==5.0.0",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/jometzner/ish-worktool#readme"
Issues = "https://github.com/jometzner/ish-worktool/issues"
Source = "https://github.com/jometzner/ish-worktool"

[project.scripts]
wt = "workTools:WorkTools"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "dirty-tag"
